refactor: move tokens.js to inter-protocol
