+++ packages/inter-protocol/src/proposals/core-proposal.js
@@ -1,6 +1,6 @@
 // @jessie-check
 
-import { Stable } from '@agoric/vats/src/tokens.js';
+import { Stable } from '../tokens.js';
 import * as econBehaviors from './econ-behaviors.js';
 import { ECON_COMMITTEE_MANIFEST } from './startEconCommittee.js';
