+++ packages/inter-protocol/src/proposals/econ-behaviors.js
@@ -8,9 +8,9 @@ import { deeplyFulfilledObject, makeTracer } from '@agoric/internal';
 import { makeStorageNodeChild } from '@agoric/internal/src/lib-chainStorage.js';
 import '@agoric/vats/exported.js';
 import '@agoric/vats/src/core/types.js';
-import { Stable } from '@agoric/vats/src/tokens.js';
 import { E } from '@endo/far';
 import { makeScalarBigMapStore } from '@agoric/vat-data';
+import { Stable } from '../tokens.js';
 import { makeGovernedTerms as makeGovernedATerms } from '../auction/params.js';
 import { makeReserveTerms } from '../reserve/params.js';
 import { makeGovernedTerms as makeGovernedVFTerms } from '../vaultFactory/params.js';
