+++ packages/inter-protocol/test/psm/test-psm.js
@@ -28,11 +28,11 @@ import {
   produceDiagnostics,
   produceStartUpgradable,
 } from '@agoric/vats/src/core/basic-behaviors.js';
-import { Stable } from '@agoric/vats/src/tokens.js';
 import { E, Far } from '@endo/far';
 import path from 'path';
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { heapZone } from '@agoric/zone';
+import { Stable } from '../../src/tokens.js';
 import { makeAnchorAsset, startPSM } from '../../src/proposals/startPSM.js';
 import {
   makeMockChainStorageRoot,
