+++ packages/inter-protocol/test/smartWallet/boot-psm.js
@@ -33,9 +33,9 @@ import {
   WALLET_FACTORY_MANIFEST,
 } from '@agoric/vats/src/core/startWalletFactory.js';
 import * as utils from '@agoric/vats/src/core/utils.js';
-import { Stable, Stake } from '@agoric/vats/src/tokens.js';
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { heapZone } from '@agoric/zone';
+import { Stable, Stake } from '../../src/tokens.js';
 import {
   ECON_COMMITTEE_MANIFEST,
   startEconomicCommittee,
