+++ packages/solo/package.json
@@ -26,6 +26,7 @@
     "@agoric/assert": "^0.6.0",
     "@agoric/cache": "^0.3.2",
     "@agoric/cosmic-swingset": "^0.41.3",
+    "@agoric/inter-protocol": "^0.16.1",
     "@agoric/internal": "^0.3.2",
     "@agoric/notifier": "^0.6.2",
     "@agoric/spawner": "^0.6.8",
