+++ packages/solo/test/test-home.js
@@ -5,9 +5,9 @@ import { test } from '@agoric/swingset-vat/tools/prepare-test-env-ava.js';
 import bundleSourceAmbient from '@endo/bundle-source';
 import { AmountMath } from '@agoric/ertp';
 import { TimeMath } from '@agoric/time';
-import { Stable } from '@agoric/vats/src/tokens.js';
 import { Far } from '@endo/marshal';
 import { resolve as importMetaResolve } from 'import-meta-resolve';
+import { Stable } from '@agoric/inter-protocol/src/tokens.js';
 
 import { makeFixture, E } from './captp-fixture.js';
