+++ packages/vats/src/core/types.js
@@ -131,7 +131,7 @@
  */
 
 /**
- * @typedef {import('../tokens.js').TokenKeyword} TokenKeyword
+ * @typedef {import('@agoric/inter-protocol/src/tokens.js').TokenKeyword} TokenKeyword
  *
  * @typedef {{
  *   issuer: |
