+++ packages/vats/test/test-demoAMM.js
@@ -1,7 +1,7 @@
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { test } from '@agoric/swingset-vat/tools/prepare-test-env-ava.js';
 import { AmountMath, makeIssuerKit } from '@agoric/ertp';
-import { Stable } from '../src/tokens.js';
+import { Stable } from '@agoric/inter-protocol/src/tokens.js';
 
 import {
   AMMDemoState,
